api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
anthropic>=0.40.0
stripe>=11.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
deepgram-sdk>=3.0.0
slowapi>=0.1.9
structlog>=24.0.0
pydantic-ai[anthropic]>=0.1.0
inngest>=0.4.0
PyJWT[crypto]>=2.8.0
